requirements/base.txt
django
djangorestframework
python-dotenv
psycopg2-binary
django-environ
drf-spectacular
factory-boy
